# api/swaggerV2.yaml
...
swagger: "2.0"
info:
  description: >-
    <div class="postman-run-button"
      data-postman-action="collection/import"
      data-postman-var-1="4cc14a96eab965c49970"
      data-postman-param="env%5Blocal%5D=W3sidHlwZSI6InRleHQiLCJlbmFibGVkIjp0cnVlLCJrZXkiOiJzdXBlcnZpc29yLmhvc3QiLCJ2YWx1ZSI6ImxvY2FsaG9zdDo1MDAwIiwiaG92ZXJlZCI6ZmFsc2V9LHsidHlwZSI6InRleHQiLCJlbmFibGVkIjp0cnVlLCJrZXkiOiJtb25pdG9yLmhvc3QiLCJ2YWx1ZSI6ImxvY2FsaG9zdDo1MDAxIiwiaG92ZXJlZCI6ZmFsc2V9LHsidHlwZSI6InRleHQiLCJlbmFibGVkIjp0cnVlLCJrZXkiOiJnb3Zlcm5pZnktcmVnaXN0cnkuaG9zdCIsInZhbHVlIjoibG9jYWxob3N0OjgwODAiLCJob3ZlcmVkIjpmYWxzZX0seyJrZXkiOiJnb3Zlcm5pZnktY29tcHV0ZXItcHBpbm90Lmhvc3QiLCJ2YWx1ZSI6ImxvY2FsaG9zdDo4MTgxIiwidHlwZSI6InRleHQiLCJlbmFibGVkIjp0cnVlLCJob3ZlcmVkIjpmYWxzZX1d"></div>
    <script type="text/javascript">
      (function (p,o,s,t,m,a,n) {
        !p[s] && (p[s] = function () { (p[t] || (p[t] = [])).push(arguments); });
        !o.getElementById(s+t) && o.getElementsByTagName("head")[0].appendChild((
          (n = o.createElement("script")),
          (n.id = s+t), (n.async = 1), (n.src = m), n
        ));
      }(window, document, "_pm", "PostmanRunObject", "https://run.pstmn.io/button.js"));
    </script>
  version: "2.0.0"
  title: "Governify Registry"
  termsOfService: "www.isa.us.es"
  contact:
    name: "ISA Research Group"
    url: "www.isa.us.es"
    email: "isagroup.us@gmail.com"
  license:
    name: "MIT"
    url: "http://opensource.org/licenses/MIT"
host: "registry.governify.io"
basePath: "/api/v2"
...
schemes:
...
consumes:
...
produces:
...
paths:
...
  /states/{agreement}:
    get:
      tags:
      - "state-registry"
      description: "XXX"
      operationId: "statesAgreementGET"
      parameters:
      - name: "agreement"
        in: "path"
        description: "XXX"
        required: true
        type: string
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/agreement_state"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/inline_response_400"
      x-swagger-router-controller: "StateRegistry"

    # DELETE ALL STATES OF AN AGREEMENTS
    delete:
      tags:
      - "state-registry"
      description: "XXX"
      operationId: "statesAgreementDELETE"
      parameters:
      - name: "agreement"
        in: "path"
        description: "XXX"
        required: true
        type: string
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/inline_response_400"
      x-swagger-router-controller: "StateRegistry"

  # RELOAD STATES OF AN AGREEMENTS
...
definitions:
...
  agreement_state:
    type: object
...
  inline_response_400:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
...